+++ old/.gitignore
@@ -8,3 +8,9 @@ Cargo.lock
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
+
+static/css
+static/js
+
+node_modules
+pnpm-lock.yaml
