+++ old/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "baski"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rocket = "0.4.11"
+serde = "1.0.152"
+serde_derive = "1.0.152"
+serde_json = "1.0.91"
+grass = "0.12.2"
+notify = "4.0.17"
+
+[dependencies.rocket_contrib]
+version = "0.4.11"
+features = ["tera_templates"]
