+++ old/src/routes/static_files.rs
