+++ old/src/routes/templates.rs
